refactor(d3d11): decouple device context from command queue
